feat(LHUB005): move service package and add coverage exclusions
