+++ sonar-project.properties
@@ -13,18 +13,13 @@ sonar.coverage.jacoco.xmlReportPaths=target/site/jacoco/jacoco.xml
 
 sonar.tests=src/test/java/
 
-sonar.exclusions=\
-  src/main/java/com/godev/linkhubservice/LinkHubServiceApplication.java,\
-  **/domain/**
-
 sonar.coverage.exclusions=\
   src/main/java/com/godev/linkhubservice/LinkHubServiceApplication.java,\
-  **/domain/**
+  src/main/java/com/godev/linkhubservice/domain/exceptions/*.java,\
+  **/domain/**,\
+  **/properties/**,\
+  **/security/**
+
 
 sonar.test.inclusions=\
-  src/main/java/com/godev/linkhubservice/LinkHubServiceApplication.java,\
-  **/domain/**
-
-sonar.cpd.exclusions=\
-  src/main/java/com/godev/linkhubservice/LinkHubServiceApplication.java,\
-  **/domain/**
+  src/main/java/com/godev/linkhubservice/LinkHubServiceApplication.java
