+++ src/main/java/com/godev/linkhubservice/services/AccountService.java
@@ -1,4 +1,4 @@
-package com.godev.linkhubservice.domain.services;
+package com.godev.linkhubservice.services;
 
 import com.godev.linkhubservice.domain.vo.AccountRequest;
 import com.godev.linkhubservice.domain.vo.AccountResponse;
