+++ src/test/java/com/godev/linkhubservice/domain/services/impl/AccountServiceImplTest.java
@@ -6,6 +6,7 @@ import com.godev.linkhubservice.domain.repository.AccountRepository;
 import com.godev.linkhubservice.helpers.AccountMockBuilder;
 import com.godev.linkhubservice.helpers.AccountRequestMockBuilder;
 import com.godev.linkhubservice.helpers.AuthRequestMockBuilder;
+import com.godev.linkhubservice.services.impl.AccountServiceImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
